Do not get properties from cache WE-274
The information required to construct the uri is already available to
us so don't hit the in-memory x-document to get this info.

This does raise a separate issue. If the XDocument object should have
protected access then the DynamicImageCacheManager.Exists() method is
currently not thread safe. However, adding a lock to this will just move
the bottleneck that we're currently experiencing on our services.

We should look optimising our cache performance further.